pyproject.toml
[tool.poetry]
name = "cubyc"
version = "0.1.0"
description = "The repository for all your experiments"
authors = [
    "Esteban Safranchik <esteban@cubyc.com>",
    "Jens C. Rischbieth <jens@cubyc.com>"
]
license = "LGPL-3.0"
repository = "https://github.com/cubyc-dev/cubyc"
readme = "README.md"
documentation = "https://docs.cubyc.com"
classifiers = [
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: SQL",
    "License :: OSI Approved :: GNU Lesser General Public License v3 (LGPLv3)"
]

[tool.poetry.dependencies]
python = "^3.8"
aiohttp = "^3.9.5"
fasteners = "^0.19"
gitpython = "^3.1.43"
ijson = "^3.2.3"
ipython = "^7.31.1"
keyring = "^25.2.1"
pandas = "^1.3.0"
psutil = "^5.9.8"
pydantic = "^2.7.1"
pyyaml = "^6.0.1"
rich = "^12.4.0"
tabulate = "^0.8.9"
typer = "^0.12.0"
wonderwords = "^2.2.0"

numpy = [
    { version = "^1.10.0", python = "<=3.11" },
    { version = "^1.20.0", python = ">=3.12" }
]

duckdb = [
    { version = "^0.8.0", python = "<3.12" },
    { version = "^1.0.0", python = ">=3.12" }
]

[tool.poetry.scripts]
cubyc = "cubyc.cli:app"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"